Switch to custom 'committers' role for new repos
This switches the default repository rights for committers from 'push' to 'committers', which is a custom role that grants the same rights as 'push' (write) plus the ability to create protected tags when tag rules are in place. See https://issues.apache.org/jira/browse/INFRA-24644 for additional context on this role.